Ensure the DHCP leases file is removed
On Debian 10 and others a machine specific identifier is generated and
written to the DHCP leases file. When requesting or renewing an address,
the identifier is sent to the DHCP server. This allows the server to
uniquely identify each client.
At shutdown, ifdown/dhclient will recreate the leases file if it is
missing. As such, to ensure the leases file is removed, we need to wait
for ifdown/dhclient to complete before attempting to remove the leases
file.